refac: 메인 함수의 기능별 메서드 분리
